third_party/blink/web_tests/external/wpt/css/css-shapes/shape-outside/values/shape-outside-ellipse-001-expected.txt

This is a testharness.js-based test.
FAIL ellipse() - inline assert_equals: expected "ellipse(at 50% 50%)" but got "ellipse()"
FAIL ellipse(50px 100px) - inline assert_equals: expected "ellipse(50px 100px at 50% 50%)" but got "ellipse(50px 100px)"
FAIL ellipse(100px 100px) - inline assert_equals: expected "ellipse(100px 100px at 50% 50%)" but got "ellipse(100px 100px)"
FAIL ellipse(25% 50%) - inline assert_equals: expected "ellipse(25% 50% at 50% 50%)" but got "ellipse(25% 50%)"
FAIL ellipse(50px 25%) - inline assert_equals: expected "ellipse(50px 25% at 50% 50%)" but got "ellipse(50px 25%)"
FAIL ellipse(25% 50px) - inline assert_equals: expected "ellipse(25% 50px at 50% 50%)" but got "ellipse(25% 50px)"
FAIL ellipse(25% closest-side) - inline assert_equals: expected "ellipse(25% closest-side at 50% 50%)" but got "ellipse(25% closest-side)"
FAIL ellipse(25px closest-side) - inline assert_equals: expected "ellipse(25px closest-side at 50% 50%)" but got "ellipse(25px closest-side)"
FAIL ellipse(closest-side 75%) - inline assert_equals: expected "ellipse(closest-side 75% at 50% 50%)" but got "ellipse(closest-side 75%)"
FAIL ellipse(closest-side 75px) - inline assert_equals: expected "ellipse(closest-side 75px at 50% 50%)" but got "ellipse(closest-side 75px)"
FAIL ellipse(25% farthest-side) - inline assert_equals: expected "ellipse(25% farthest-side at 50% 50%)" but got "ellipse(25% farthest-side)"
FAIL ellipse(25px farthest-side) - inline assert_equals: expected "ellipse(25px farthest-side at 50% 50%)" but got "ellipse(25px farthest-side)"
FAIL ellipse(farthest-side 75%) - inline assert_equals: expected "ellipse(farthest-side 75% at 50% 50%)" but got "ellipse(farthest-side 75%)"
FAIL ellipse(farthest-side 75px) - inline assert_equals: expected "ellipse(farthest-side 75px at 50% 50%)" but got "ellipse(farthest-side 75px)"
FAIL ellipse(closest-side closest-side) - inline assert_equals: expected "ellipse(at 50% 50%)" but got "ellipse()"
FAIL ellipse(farthest-side farthest-side) - inline assert_equals: expected "ellipse(farthest-side farthest-side at 50% 50%)" but got "ellipse(farthest-side farthest-side)"
FAIL ellipse(closest-side farthest-side) - inline assert_equals: expected "ellipse(closest-side farthest-side at 50% 50%)" but got "ellipse(closest-side farthest-side)"
FAIL ellipse(farthest-side closest-side) - inline assert_equals: expected "ellipse(farthest-side closest-side at 50% 50%)" but got "ellipse(farthest-side closest-side)"
FAIL ellipse() - computed assert_equals: expected "ellipse(at 50% 50%)" but got "ellipse()"
FAIL ellipse(50px 100px) - computed assert_equals: expected "ellipse(50px 100px at 50% 50%)" but got "ellipse(50px 100px)"
FAIL ellipse(100px 100px) - computed assert_equals: expected "ellipse(100px 100px at 50% 50%)" but got "ellipse(100px 100px)"
FAIL ellipse(25% 50%) - computed assert_equals: expected "ellipse(25% 50% at 50% 50%)" but got "ellipse(25% 50%)"
FAIL ellipse(50px 25%) - computed assert_equals: expected "ellipse(50px 25% at 50% 50%)" but got "ellipse(50px 25%)"
FAIL ellipse(25% 50px) - computed assert_equals: expected "ellipse(25% 50px at 50% 50%)" but got "ellipse(25% 50px)"
FAIL ellipse(25% closest-side) - computed assert_equals: expected "ellipse(25% closest-side at 50% 50%)" but got "ellipse(25% closest-side)"
FAIL ellipse(25px closest-side) - computed assert_equals: expected "ellipse(25px closest-side at 50% 50%)" but got "ellipse(25px closest-side)"
FAIL ellipse(closest-side 75%) - computed assert_equals: expected "ellipse(closest-side 75% at 50% 50%)" but got "ellipse(closest-side 75%)"
FAIL ellipse(closest-side 75px) - computed assert_equals: expected "ellipse(closest-side 75px at 50% 50%)" but got "ellipse(closest-side 75px)"
FAIL ellipse(25% farthest-side) - computed assert_equals: expected "ellipse(25% farthest-side at 50% 50%)" but got "ellipse(25% farthest-side)"
FAIL ellipse(25px farthest-side) - computed assert_equals: expected "ellipse(25px farthest-side at 50% 50%)" but got "ellipse(25px farthest-side)"
FAIL ellipse(farthest-side 75%) - computed assert_equals: expected "ellipse(farthest-side 75% at 50% 50%)" but got "ellipse(farthest-side 75%)"
FAIL ellipse(farthest-side 75px) - computed assert_equals: expected "ellipse(farthest-side 75px at 50% 50%)" but got "ellipse(farthest-side 75px)"
FAIL ellipse(closest-side closest-side) - computed assert_equals: expected "ellipse(at 50% 50%)" but got "ellipse()"
FAIL ellipse(farthest-side farthest-side) - computed assert_equals: expected "ellipse(farthest-side farthest-side at 50% 50%)" but got "ellipse(farthest-side farthest-side)"
FAIL ellipse(closest-side farthest-side) - computed assert_equals: expected "ellipse(closest-side farthest-side at 50% 50%)" but got "ellipse(closest-side farthest-side)"
FAIL ellipse(farthest-side closest-side) - computed assert_equals: expected "ellipse(farthest-side closest-side at 50% 50%)" but got "ellipse(farthest-side closest-side)"
Harness: the test ran to completion.
